fix(compare): reject evidence flags on release fan-out; keep cli.py under cap
- Reject inline build/source evidence flags (--max/--depth/--old-new-sources/
  --old-new-build-info) on directory/package (release) compares instead of
  silently dropping them — the per-library fan-out forwards only release kwargs
  (Codex review). Parametrized regression test added.
- Move the set-input rejection guards (compile-context + evidence) and
  _config_has_compile_block from cli.py to the leaf cli_resolve.py, keeping
  cli.py under the 2000-line hard cap (the new guard pushed it to 2038).
- Mark dump_cmd's _resolved_collect_mode branch (only reached via compare's
  inline embed, integration) with a coverage pragma.
- Update the published Action guidance: `depth: graph` → `source-method: s4`
  (the removed graph rung), per the clean-removal directive.
